# backend/internal/api/handlers/files.go
package handlers

import (
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/google/uuid"
)

// GetTaskFiles returns all files generated by the LLM during a specific task.
// GET /api/tasks/{id}/files
func (h *Handlers) GetTaskFiles(w http.ResponseWriter, r *http.Request) {
	if h.GeneratedFiles == nil {
		writeJSON(w, http.StatusOK, []struct{}{})
		return
	}

	rawID := chi.URLParam(r, "id")
	taskID, err := uuid.Parse(rawID)
	if err != nil {
		writeError(w, http.StatusBadRequest, "invalid task id")
		return
	}

	files, err := h.GeneratedFiles.ListByTask(r.Context(), taskID)
	if err != nil {
		writeError(w, http.StatusInternalServerError, "failed to list files")
		return
	}

	writeJSON(w, http.StatusOK, files)
}